Move resend NYM test into separate file.
